Begun Removing RoundImageEx code from Branch
This reverts commit c8beee42242b73e0877ad875462709957ee8d89c.

# Conflicts:
#	Microsoft.Toolkit.Uwp.UI.Controls/Themes/Generic.xaml
