+++ Microsoft.Toolkit.Uwp.UI.Controls/ImageEx/ImageEx.Placeholder.cs
@@ -11,15 +11,16 @@
 // ******************************************************************
 
 using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Media;
 
 namespace Microsoft.Toolkit.Uwp.UI.Controls
 {
     /// <summary>
-    /// Shared Code for ImageEx and RoundImageEx
+    /// The ImageEx control extends the default Image platform control improving the performance and responsiveness of your Apps.
+    /// Source images are downloaded asynchronously showing a load indicator while in progress.
+    /// Once downloaded, the source image is stored in the App local cache to preserve resources and load time next time the image needs to be displayed.
     /// </summary>
-    public partial class ImageExBase
+    public partial class ImageEx
     {
         /// <summary>
         /// Identifies the <see cref="PlaceholderSource"/> dependency property.
@@ -27,7 +28,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         public static readonly DependencyProperty PlaceholderSourceProperty = DependencyProperty.Register(
             nameof(PlaceholderSource),
             typeof(ImageSource),
-            typeof(ImageExBase),
+            typeof(ImageEx),
             new PropertyMetadata(default(ImageSource)));
 
         /// <summary>
@@ -36,7 +37,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         public static readonly DependencyProperty PlaceholderStretchProperty = DependencyProperty.Register(
             nameof(PlaceholderStretch),
             typeof(Stretch),
-            typeof(ImageExBase),
+            typeof(ImageEx),
             new PropertyMetadata(default(Stretch)));
 
         /// <summary>
@@ -49,31 +50,6 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
         {
             get { return (ImageSource)GetValue(PlaceholderSourceProperty); }
             set { SetValue(PlaceholderSourceProperty, value); }
-        }
-
-        private static void PlaceholderSourceChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            var control = d as ImageExBase;
-
-            if (e.OldValue == null || e.NewValue == null || !e.OldValue.Equals(e.NewValue))
-            {
-                if (control is RoundImageEx round)
-                {
-                    if (round.ShowPlaceholderStroke)
-                    {
-                        if (e.NewValue == null)
-                        {
-                            VisualStateManager.GoToState(round, StrokeUnloaded, true);
-                        }
-                        else
-                        {
-                            VisualStateManager.GoToState(round, ShowStrokeState, true);
-                        }
-                    }
-                }
-
-                control?.SetSource(e.NewValue);
-            }
         }
 
         /// <summary>
