+++ Microsoft.Toolkit.Uwp.UI.Controls/ImageEx/ImageEx.cs
@@ -14,7 +14,6 @@ using System;
 using Windows.Foundation;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Media;
 
 namespace Microsoft.Toolkit.Uwp.UI.Controls
 {
@@ -23,15 +22,86 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
     /// Source images are downloaded asynchronously showing a load indicator while in progress.
     /// Once downloaded, the source image is stored in the App local cache to preserve resources and load time next time the image needs to be displayed.
     /// </summary>
-    public partial class ImageEx : ImageExBase
+    [TemplateVisualState(Name = LoadingState, GroupName = CommonGroup)]
+    [TemplateVisualState(Name = LoadedState, GroupName = CommonGroup)]
+    [TemplateVisualState(Name = UnloadedState, GroupName = CommonGroup)]
+    [TemplateVisualState(Name = FailedState, GroupName = CommonGroup)]
+    [TemplatePart(Name = PartImage, Type = typeof(Image))]
+    [TemplatePart(Name = PartProgress, Type = typeof(ProgressRing))]
+    public partial class ImageEx : Control
     {
+        private const string PartImage = "Image";
+        private const string PartProgress = "Progress";
+        private const string CommonGroup = "CommonStates";
+        private const string LoadingState = "Loading";
+        private const string LoadedState = "Loaded";
+        private const string UnloadedState = "Unloaded";
+        private const string FailedState = "Failed";
+
+        private Image _image;
+        private ProgressRing _progress;
+        private bool _isInitialized;
+
         /// <summary>
         /// Initializes a new instance of the <see cref="ImageEx"/> class.
         /// </summary>
         public ImageEx()
-            : base()
         {
             DefaultStyleKey = typeof(ImageEx);
+        }
+
+        /// <summary>
+        /// Update the visual state of the control when its template is changed.
+        /// </summary>
+        protected override void OnApplyTemplate()
+        {
+            if (_image != null)
+            {
+                _image.ImageOpened -= OnImageOpened;
+                _image.ImageFailed -= OnImageFailed;
+            }
+
+            _image = GetTemplateChild(PartImage) as Image;
+            _progress = GetTemplateChild(PartProgress) as ProgressRing;
+
+            _isInitialized = true;
+
+            SetSource(Source);
+
+            if (_image != null)
+            {
+                _image.ImageOpened += OnImageOpened;
+                _image.ImageFailed += OnImageFailed;
+            }
+
+            base.OnApplyTemplate();
+        }
+
+        /// <inheritdoc/>
+        protected override Size ArrangeOverride(Size finalSize)
+        {
+            var newSquareSize = Math.Min(finalSize.Width, finalSize.Height) / 8.0;
+
+            if (_progress?.Width == newSquareSize)
+            {
+                _progress.Height = newSquareSize;
+            }
+
+            return base.ArrangeOverride(finalSize);
+        }
+
+        private void OnImageOpened(object sender, RoutedEventArgs e)
+        {
+            ImageOpened?.Invoke(this, e);
+            ImageExOpened?.Invoke(this, new ImageExOpenedEventArgs());
+            VisualStateManager.GoToState(this, LoadedState, true);
+        }
+
+        private void OnImageFailed(object sender, ExceptionRoutedEventArgs e)
+        {
+            ImageFailed?.Invoke(this, e);
+            ImageExFailed?.Invoke(this, new ImageExFailedEventArgs(new Exception(e.ErrorMessage)));
+            VisualStateManager.GoToState(this, FailedState, true);
         }
     }
 }
