+++ Microsoft.Toolkit.Uwp.UI.Controls/ImageEx/ImageEx.xaml
