--- Microsoft.Toolkit.Uwp.UI.Controls/ImageEx/ImageExBase.Members.cs
@@ -1,101 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Media.Imaging;
-
-namespace Microsoft.Toolkit.Uwp.UI.Controls
-{
-    /// <summary>
-    /// Shared Code for ImageEx and RoundImageEx
-    /// </summary>
-    public partial class ImageExBase
-    {
-        /// <summary>
-        /// Identifies the <see cref="Stretch"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty StretchProperty = DependencyProperty.Register(nameof(Stretch), typeof(Stretch), typeof(ImageExBase), new PropertyMetadata(Stretch.Uniform));
-
-        /// <summary>
-        /// Identifies the <see cref="DecodePixelHeight"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty DecodePixelHeightProperty = DependencyProperty.Register(nameof(DecodePixelHeight), typeof(int), typeof(ImageExBase), new PropertyMetadata(0));
-
-        /// <summary>
-        /// Identifies the <see cref="DecodePixelType"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty DecodePixelTypeProperty = DependencyProperty.Register(nameof(DecodePixelType), typeof(int), typeof(ImageExBase), new PropertyMetadata(DecodePixelType.Physical));
-
-        /// <summary>
-        /// Identifies the <see cref="DecodePixelWidth"/> dependency property.
-        /// </summary>
-        public static readonly DependencyProperty DecodePixelWidthProperty = DependencyProperty.Register(nameof(DecodePixelWidth), typeof(int), typeof(ImageExBase), new PropertyMetadata(0));
-
-        /// <summary>
-        /// Event raised if the image failed loading.
-        /// </summary>
-        [Obsolete("This event is obsolete; use ImageExFailed event instead")]
-        public event ExceptionRoutedEventHandler ImageFailed;
-
-        /// <summary>
-        /// Event raised when the image is successfully loaded and opened.
-        /// </summary>
-        [Obsolete("This event is obsolete; use ImageExOpened event instead")]
-        public event RoutedEventHandler ImageOpened;
-
-        /// <summary>
-        /// Event raised if the image failed loading.
-        /// </summary>
-        public event ImageExFailedEventHandler ImageExFailed;
-
-        /// <summary>
-        /// Event raised when the image is successfully loaded and opened.
-        /// </summary>
-        public event ImageExOpenedEventHandler ImageExOpened;
-
-        /// <summary>
-        /// Gets or sets DecodePixelHeight for underlying bitmap
-        /// </summary>
-        public int DecodePixelHeight
-        {
-            get { return (int)GetValue(DecodePixelHeightProperty); }
-            set { SetValue(DecodePixelHeightProperty, value); }
-        }
-
-        /// <summary>
-        /// Gets or sets DecodePixelType for underlying bitmap
-        /// </summary>
-        public DecodePixelType DecodePixelType
-        {
-            get { return (DecodePixelType)GetValue(DecodePixelTypeProperty); }
-            set { SetValue(DecodePixelTypeProperty, value); }
-        }
-
-        /// <summary>
-        /// Gets or sets DecodePixelWidth for underlying bitmap
-        /// </summary>
-        public int DecodePixelWidth
-        {
-            get { return (int)GetValue(DecodePixelWidthProperty); }
-            set { SetValue(DecodePixelWidthProperty, value); }
-        }
-
-        public Stretch Stretch
-        {
-            get { return (Stretch)GetValue(StretchProperty); }
-            set { SetValue(StretchProperty, value); }
-        }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether gets or sets cache state
-        /// </summary>
-        public bool IsCacheEnabled
-        {
-            get; set;
-        }
-    }
-}
