+++ Microsoft.Toolkit.Uwp.UI.Controls/Microsoft.Toolkit.Uwp.UI.Controls.csproj
@@ -72,12 +72,6 @@
     <Compile Include="HamburgerMenu\MenuItems\HamburgerMenuImageItem.cs" />
     <Compile Include="HeaderedTextBlock\HeaderedTextBlock.cs" />
     <Compile Include="HeaderedTextBlock\HeaderedTextBlock.Properties.cs" />
-    <Compile Include="ImageEx\ImageExBase.Placeholder.cs" />
-    <Compile Include="ImageEx\ImageExBase.Source.cs" />
-    <Compile Include="ImageEx\ImageExBase.cs" />
-    <Compile Include="ImageEx\ImageExBase.Members.cs" />
-    <Compile Include="ImageEx\RoundImageEx.cs" />
-    <Compile Include="ImageEx\RoundImageEx.Members.cs" />
     <Compile Include="MarkdownTextBlock\Display\ILinkRegister.cs" />
     <Compile Include="MarkdownTextBlock\Display\MarkdownTable.cs" />
     <Compile Include="MarkdownTextBlock\Display\XamlRenderer.cs" />
@@ -152,6 +146,8 @@
     <Compile Include="ControlHelpers.Math.cs" />
     <Compile Include="ImageEx\ImageEx.cs" />
     <Compile Include="ImageEx\ImageEx.Members.cs" />
+    <Compile Include="ImageEx\ImageEx.Placeholder.cs" />
+    <Compile Include="ImageEx\ImageEx.Source.cs" />
     <Compile Include="ScrollHeader\ScrollHeader.cs" />
     <Compile Include="ImageEx\ImageExFailedEventArgs.cs" />
     <Compile Include="ImageEx\ImageExOpenedEventArgs.cs" />
@@ -200,11 +196,6 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Page>
     <Page Include="ImageEx\ImageEx.xaml">
-      <Generator>MSBuild:Compile</Generator>
-      <SubType>Designer</SubType>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Page>
-    <Page Include="ImageEx\RoundImageEx.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
